Don't use vim-pencil automatically
It doesn't seem to work terribly well when trying to insert lists in
markdown, as it keeps trying to join the lines together as if you're
writing a paragraph of prose. So turning this off for now, I can turn it
on manually with `:Pencil` if/when I want to use it.